legacy/backend/pyproject.toml
[project]
name = "axion"
version = "1.0.0"
description = "Axion — The Operating System for the Autonomous Enterprise"
authors = [
    { name = "Axion Labs SAS", email = "engineering@axion.ai" }
]
license = { text = "Proprietary" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.28.0",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "mistralai>=1.2.0",
    "tiktoken>=0.8.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.17",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "prometheus-client>=0.21.0",
    "stripe>=11.4.0",
    "boto3>=1.35.0",
    "pgvector>=0.3.0",
    "networkx>=3.4.0",
    "tenacity>=9.0.0",
    "orjson>=3.10.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "faker>=33.0.0",
    "freezegun>=1.5.0",
]

[project.scripts]
axion-api = "axion.main:run"
axion-worker = "axion.workers.celery_app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/axion"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
